pip-requirements.txt
Django==1.4.2
MySQL-python==1.2.4c1
PIL==1.1.7
South==0.7.6
argparse==1.2.1
cmsplugin-plaintext==0.1.1-alpha-2
django-classy-tags==0.3.4.1
django-cms==2.3.4
django-mptt==0.5.2
django-rosetta==0.6.8
django-sekizai==0.6.1
gunicorn==0.13.4
html5lib==0.95
wsgiref==0.1.2
django-translatable==0.1.1
